step2: create ResultView and delegate function
