+++ src/main/kotlin/blackjack/Main.kt
@@ -3,10 +3,18 @@ package blackjack
 fun main() {
     val cardDeck = CardDeck()
     val players = InputView.shareCard(cardDeck)
+    for (player in players) ResultView.showPlayerCards(player)
 
     for (player in players) {
-        InputView.hitCard(player, cardDeck)
+        var isHitCalled = InputView.isHitCalled(player)
+        while (isHitCalled) {
+            player.hit(cardDeck.drawCard())
+            ResultView.showPlayerCards(player)
+            if (player.isBusted) break
+
+            isHitCalled = InputView.isHitCalled(player)
+        }
     }
 
-    InputView.showResult(players)
+    ResultView.showResult(players)
 }
